kconfig: Add a configuration file for nospl
This file indicates which options are not for use in U-Boot SPL. It allows
the kconfig tool to correctly generate its output.

It was generated with

   ./tools/moveconfig.py --scan-source -u

Note that this only produces sensible results when used with the current
source code, as of this patch. Following patches adjust the code, e.g. to
drop use of CONFIG_IS_ENABLED(), so the program will not produce anything
useful when run on that version of the code.

Series-changes: 3
- Add DFU_VIRT FSL_ISBC_KEY_EXT GENERATE_SMBIOS_TABLE
  GMAC_ROCKCHIP I8259_PIC IMX_RDC NXP_FSPI PCI_PNP PHY_CADENCE_TORRENT
  QFW_PIO SEC_FIRMWARE_ARMV8_PSCI SPI_FLASH_BAR SPI_FLASH_MACRONIX
- Add CMD_BOOTEFI_BOOTMGR CMD_EFICONFIG CMD_PXE
- Fix move_config typo
- Add ARCH_VERSAL_NET CMD_DHCP ENV_WRITEABLE_LIST EXPO
- Add TARGET_KMCOGE5NE TARGET_KMETER1

Series-changes: 4
- Add the following to conf_nospl instead of creating def_bool options:
  ACPIGEN AVB_VERIFY BOOTDEV_ETH BOOTDEV_SPI_FLASH BOOTMETH_DISTRO
  BOOTMETH_DISTRO_PXE BOOTMETH_EFILOADER BOOTMETH_SANDBOX
  BOOTMETH_SCRIPT BOOTMETH_VBE_SIMPLE_OS BOOTSTD_FULL BUTTON CMDLINE
  CONSOLE_MUX CRC32_VERIFY DEVRES DM_HWSPINLOCK EFI_DEVICE_PATH_TO_TEXT
  EFI_LOADER ERRNO_STR EVENT EVENT_DYNAMIC FIT_CIPHER FSP_VERSION2
  GENERATE_ACPI_TABLE HUSH_PARSER I2C_CROS_EC_LDO I2C_CROS_EC_TUNNEL
  IOMMU OF_LIVE PINCTRL_ARMADA_38X PXE_UTILS QCOM_PMIC_GPIO
  RESET_MEDIATEK RESET_ROCKCHIP SYS_CONSOLE_IS_IN_ENV UT_COMPRESSION
  VIRTIO

Signed-off-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit 50133420ad1356abe4e2e4610b509ff093bd802d)