freescale: Fix odd use of ESDHCI_QUIRK_BROKEN_TIMEOUT_VALUE
In fsl_esdhc.c this is not a CONFIG option, so we should not be using
IS_ENABLED() on it, particularly not when it is not defined to anything,
which shows up as calling IS_ENABLED() with no arguments.

Define it to 0 at the top of the file if it is not already defined and
check it normally. This sort-of matches the behaviour in fsl_esdhc_imx.c
which had a similar problem.

Signed-off-by: Simon Glass <sjg@chromium.org>