rockchip: boot: Correct use of OF_LIVE
This is never enabled in SPL so it should be checked using
CONFIG_IS_ENABLED(). But really it should not be checked at all, since
the ofnode API handles this automatically.

Adjust the strange code in the rockchip pinctrl driver and a Makefile
rule, so that split config will work correctly.

Signed-off-by: Simon Glass <sjg@chromium.org>
(cherry picked from commit 3cf213237a131a9f8be53818625429f5d4d81678)