video: nexell: unsigned parameter cannot be negative
The parameter 'alpha' is declared as an unsigned type so cannot be
negative. The code to test it as being less than zero will always fail
and so is redundant and should be removed.

This issue was found by Smatch.

Signed-off-by: Andrew Goodbody <andrew.goodbody@linaro.org>
Reviewed-by: Stefan Bosch <stefan_b@posteo.net>
Signed-off-by: Minkyu Kang <mk7.kang@samsung.com>
(cherry picked from commit 548e9a254c7d0bd9796decfea43d7759070c7312)