pwm: make sandbox depend on DM_PWM
Since it is registered as a U_CLASS_DRIVER, Sandbox PWM driver is a
Driver Model Driver and thus to be usable depends on DM_PWM to be
selected.

Let's make sure of that via the appropriate Kconfig option.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
(cherry picked from commit 534892522514b2e72614cd0af1d352879cf7604a)